FEAT-#2197: enable pydocstyle check for series.py
